refactor : refactored component classnames and structures (folder)
