+++ src/pages/Main/ClickList/ClickList.scss
@@ -1,0 +1,92 @@
+@import '../../../styles/constant';
+
+.ClickList {
+  display: flex;
+  justify-content: space-between;
+
+  .list {
+    width: 50%;
+
+    .clickListItem {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      padding: 15px 20px;
+      border-top: 1px $border solid;
+      border-bottom: 1px $border solid;
+      opacity: 0.3;
+      transition: 200ms ease-in-out;
+      cursor: pointer;
+      margin-top: -1px;
+
+      &:first-child {
+        margin-top: 0px;
+      }
+
+      .fa-chevron-right {
+        opacity: 0;
+        pointer-events: none;
+      }
+
+      &.active {
+        opacity: 1;
+
+        & > .fa-chevron-right {
+          opacity: 1;
+          transition: 600ms ease-in-out;
+        }
+      }
+
+      &:hover {
+        opacity: 1;
+      }
+
+      .clickListItemTitle {
+        pointer-events: none;
+
+        h3 {
+          font-family: 'Source Sans Pro', sans-serif;
+          font-weight: 700;
+          font-size: 16px;
+          margin-bottom: 4px;
+          pointer-events: none;
+        }
+
+        h4 {
+          font-family: 'Cutive Mono', monospace;
+          font-size: 14px;
+          line-height: 1.2em;
+          pointer-events: none;
+        }
+      }
+    }
+  }
+
+  .clickListImageWrap {
+    width: 40%;
+    margin-right: 24px;
+    overflow: hidden;
+
+    .clickListImageBlock {
+      max-height: 0;
+      opacity: 0;
+
+      &.active {
+        max-height: 1000px;
+        opacity: 1;
+        transition: opacity 300ms ease-in-out;
+      }
+    }
+
+    .clickListImage {
+      width: 100%;
+      margin-bottom: 22px;
+    }
+
+    .clickListImgDescription {
+      font-family: 'Cutive Mono', monospace;
+      font-size: 13px;
+      line-height: 1.2em;
+    }
+  }
+}
