+++ src/pages/Main/ClickList/ClickListItem.js
@@ -1,21 +1,15 @@
 import React from 'react';
 
-const ClickListItem = ({
-  selectedState,
-  index,
-  liClassName,
-  onClick,
-  data,
-}) => {
+const ClickListItem = ({ selectedState, index, onClick, data }) => {
   return (
     <li
       className={
-        index === selectedState ? `${liClassName} active` : liClassName
+        index === selectedState ? `clickListItem active` : 'clickListItem'
       }
       index={index}
       onClick={onClick}
     >
-      <div className="breweryTitle">
+      <div className="clickListItemTitle">
         <h3>{data.brewery.toUpperCase()}</h3>
         <h4>{data.location}</h4>
       </div>
