+++ src/pages/Main/MainSection.scss
@@ -1,6 +1,6 @@
 @import '../../styles/constant';
 
-.mainSection {
+.MainSection {
   border: 1px $border solid;
   border-top: none;
   width: calc(100% - 250px);
