+++ src/pages/Main/Maincard.scss
@@ -1,6 +1,6 @@
 @import '../../styles/constant.scss';
 
-.mainCard {
+.MainCard {
   color: $text;
   text-decoration: none;
